Update README and add requirements.txt
This adds installation instructions to the README and
adds a requirements.txt file that can be used to install
the dependencies for the project.

You don't need to follow these instructions exactly, but
doing it this way will make sure the dependencies don't
conflict with other Python installations on your machine.